Module Ppx_log_types.Message_sexp

module Label : sig ... end

type t

A t represents a %message sexp with additional structured information that distinguishes the message label from the typed list of tags that follow.

val create : ?legacy_render_with_additional_parentheses:Base.bool -> Label.t Base.option -> tags:Log_tag.t Base.list -> t

val label : t -> Label.t Base.option

label is None in cases like %log log (tag1 : t1) ... or %log.global "" (tag1 : t1) ....

val tags : t -> Log_tag.t Base.list

val render : t -> Base.Sexp.t

module Unstable : sig ... end